src/main/java/lotto/model/Lottos.java
package lotto.model;

import java.util.List;
import java.util.Objects;
import java.util.function.Consumer;
import java.util.stream.Collectors;
import java.util.stream.IntStream;

public class Lottos {

    private final List<Lotto> lottos;

    public Lottos(List<Lotto> lottos) {
        this.lottos = lottos;
    }

    public int size() {
        return this.lottos.size();
    }

    public void foreach(Consumer consumer) {
        for (Lotto lotto : lottos) {
            consumer.accept(lotto);
        }
    }

    public LottoRanks match(WinningLotto winningLotto) {
        return new LottoRanks(this.lottos.stream()
                .map(o -> o.match(winningLotto))
                .collect(Collectors.toList()));
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (!(o instanceof Lottos)) return false;
        Lottos lottos1 = (Lottos) o;
        return Objects.equals(lottos, lottos1.lottos);
    }

    @Override
    public int hashCode() {
        return Objects.hashCode(lottos);
    }
}
